Delta eval test code reuse
